新增投标函数

+++ src/main/java/com/gofobao/framework/borrow/vo/request/VoBorrowByIdReq.java
@@ -9,6 +9,6 @@ import lombok.Data;
 @Data
 public class VoBorrowByIdReq {
         @ApiModelProperty("标id")
-        private Integer borrowId;
+        private Long borrowId;
 
 }
